build: use pre-generate product to distingiush product dts
Qcom kernel only support target and variant config, so add
new product config for Moto varous skus.

When a dtb or dtbo target file is only for a product, the "product" can
be specific to a product name, such as "arcf" or "ctwov". when the product
field is not specific, the dtb/dtbo file will be compiled default.

Change-Id: I9bf180331b1f87b5f595f03f9975b874b8b5d46f
Reviewed-on: https://gerrit.mot.com/2772414
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Reviewed-by: Tao Sun <suntao2@motorola.com>
Tested-by: Jira Key
Submit-Approved: Jira Key